StateReader works for cairo0 now
