+++ src/starknet/business_logic/fact_state/state.rs
@@ -1,4 +1,5 @@
 use std::collections::{HashMap, HashSet};
+use std::ops::Deref;
 
 use blockifier::execution::contract_class::{ContractClass, ContractClassV0, ContractClassV1};
 use blockifier::state::cached_state::CommitmentStateDiff;
@@ -8,6 +9,7 @@ use blockifier::test_utils::dict_state_reader::DictStateReader;
 use cairo_vm::types::errors::math_errors::MathError;
 use cairo_vm::Felt252;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
+use starknet_api::deprecated_contract_class;
 use starknet_api::hash::StarkFelt;
 use starknet_api::state::StorageKey;
 use starknet_crypto::FieldElement;
@@ -18,7 +20,7 @@ use crate::config::{
 };
 use crate::crypto::poseidon::{poseidon_hash_many_bytes, PoseidonHash};
 use crate::starknet::business_logic::fact_state::contract_class_objects::{
-    get_ffc_for_contract_class_facts, ContractClassLeaf,
+    get_ffc_for_contract_class_facts, ContractClassLeaf, DeprecatedCompiledClassFact,
 };
 use crate::starknet::business_logic::fact_state::contract_state_objects::ContractState;
 use crate::starknet::business_logic::state::state_api_objects::BlockInfo;
@@ -27,7 +29,8 @@ use crate::starkware_utils::commitment_tree::base_types::{Height, TreeIndex};
 use crate::starkware_utils::commitment_tree::binary_fact_tree::BinaryFactTree;
 use crate::starkware_utils::commitment_tree::errors::TreeError;
 use crate::starkware_utils::commitment_tree::patricia_tree::patricia_tree::PatriciaTree;
-use crate::storage::storage::{FactFetchingContext, HashFunctionType, Storage};
+use crate::storage::storage::{DbObject, FactFetchingContext, HashFunctionType, Storage};
+use crate::storage::storage_utils::deprecated_contract_class_api2vm;
 use crate::utils::{execute_coroutine, felt_api2vm, felt_vm2api};
 
 /// A class representing a combination of the onchain and offchain state.
@@ -392,28 +395,14 @@ where
     async fn get_compiled_contract_class_async(&mut self, class_hash: ClassHash) -> StateResult<ContractClass> {
         println!("SharedState as StateReader: get_compiled_contract_class {:?}", class_hash);
 
-        let storage = self.ffc_for_class_hash.acquire_storage().await;
+        let storage = self.ffc.acquire_storage().await;
 
         // first try to read as a deprecated class (no class_hash -> compiled_class_hash indirection)
-        let contract_bytes = {
-            let bytecode =
-                storage.get_value(class_hash.0.bytes()).await.map_err(|_| {
-                    StateError::StateReadError(format!("Error reading storage value for {:?}", class_hash.clone()))
-                })?;
-            bytecode
+        let deprecated_compiled_class_fact = DeprecatedCompiledClassFact::get(storage.deref(), class_hash.0.bytes()).await.unwrap();
+        // TODO: better conversion?
+        if let Some(deprecated_compiled_class_fact) = deprecated_compiled_class_fact {
+            return Ok(deprecated_contract_class_api2vm(&deprecated_compiled_class_fact.contract_definition).unwrap());
         };
-
-        if let Some(contract_bytes) = contract_bytes {
-            // we have a deprecated contract here
-            println!("Got deprecated class bytes");
-            // TODO: at this point we should not fail deserialization
-
-            // TODO: consider from_utf8_unchecked (performance improvement)
-            let contract_bytes = std::str::from_utf8(&contract_bytes).unwrap();
-
-            let contract_class = ContractClassV0::try_from_json_string(contract_bytes);
-            return Ok(ContractClass::V0(contract_class.unwrap()))
-        }
 
         // TODO: try v1 instead
         todo!("try v1");
